Initial split into requisite files - including base file
