+++ codegen/codegen.rb
@@ -4,479 +4,6 @@ require 'json'
 require 'erb'
 require 'set'
 
-# CodeGen
-# Automatic Code generation base class for generating cucumber messages
-class Codegen
-  TEMPLATES_DIRECTORY = "#{File.dirname(__FILE__)}/templates/".freeze
-
-  def initialize(paths, language_type_by_schema_type)
-    @paths = paths
-    @language_type_by_schema_type = language_type_by_schema_type
-
-    @schemas = {}
-    @enum_set = Set.new
-
-    @paths.each do |path|
-      expanded_path = File.expand_path(path)
-      begin
-        schema = JSON.parse(File.read(path))
-        add_schema(expanded_path, schema)
-      rescue JSON::ParserError => e
-        e.message << "\npath: #{path}"
-        raise e
-      end
-    end
-
-    @schemas = @schemas.sort
-    @enums = @enum_set.to_a.sort_by { |a| a[:name] }
-  end
-
-  def generate(template_name)
-    template_source = File.read("#{TEMPLATES_DIRECTORY}/#{template_name}")
-    template = ERB.new(template_source, trim_mode: '-')
-    $stdout.write template.result(binding)
-  end
-
-  private
-
-  def add_schema(key, schema)
-    @schemas[key] = schema
-    (schema['definitions'] || {}).each do |name, subschema|
-      subkey = "#{key}/#{name}"
-      add_schema(subkey, subschema)
-    end
-
-    schema['properties'].each do |property_name, property|
-      enum = property['enum']
-      if enum
-        parent_type_name = class_name(key)
-        enum_name(parent_type_name, property_name, enum)
-      end
-    end
-  end
-
-  def default_value(parent_type_name, property_name, property)
-    if property['items']
-      '[]'
-    elsif %w[string integer boolean].include?(property['type'])
-      default_value_for_non_objects(parent_type_name, property_name, property)
-    elsif property['$ref']
-      type = type_for(parent_type_name, nil, property)
-      "new #{type}()"
-    else
-      raise "Cannot create default value for #{parent_type_name}##{property.to_json}"
-    end
-  end
-
-  def default_value_for_non_objects(parent_type_name, property_name, property)
-    case property['type']
-    when 'string'
-      if property['enum']
-        enum_type_name = type_for(parent_type_name, property_name, property)
-        default_enum(enum_type_name, property)
-      else
-        "''"
-      end
-    when 'integer'; then '0'
-    when 'boolean'; then 'false'
-    end
-  end
-
-  def default_enum(enum_type_name, property)
-    "#{enum_type_name}.#{enum_constant(property['enum'][0])}"
-  end
-
-  def enum_constant(value)
-    value.gsub(/[.\/+]/, '_').upcase
-  end
-
-  def type_for(parent_type_name, property_name, property)
-    type = property['type']
-
-    if property['$ref']
-      property_type_from_ref(property['$ref'])
-    elsif type
-      property_type_from_type(parent_type_name, property_name, property)
-    else
-      # Inline schema (not supported)
-      raise "Property #{name} did not define 'type' or '$ref'"
-    end
-  end
-
-  def property_type_from_ref(ref)
-    class_name(ref)
-  end
-
-  def property_type_from_type(parent_type_name, property_name, property)
-    type = property['type']
-
-    if type == 'array'
-      array_type_for(type_for(parent_type_name, nil, property['items']))
-    else
-      unless @language_type_by_schema_type[type]
-        raise "No type mapping for JSONSchema type #{type}. Schema:\n#{JSON.pretty_generate(property)}"
-      end
-
-      if property['enum']
-        enum_type_name = enum_name(parent_type_name, property_name, property['enum'])
-        property_type_from_enum(enum_type_name)
-      else
-        @language_type_by_schema_type[type]
-      end
-    end
-  end
-
-  def property_type_from_enum(enum)
-    enum
-  end
-
-  def enum_name(parent_type_name, property_name, enum)
-    enum_type_name = "#{parent_type_name}#{capitalize(property_name)}"
-    @enum_set.add({ name: enum_type_name, values: enum })
-    enum_type_name
-  end
-
-  def class_name(ref)
-    File.basename(ref, '.json')
-  end
-
-  def capitalize(string)
-    string.sub(/./, &:upcase)
-  end
-
-  # Thank you very much rails!
-  # https://github.com/rails/rails/blob/v6.1.3.2/activesupport/lib/active_support/inflector/methods.rb#L92
-  def underscore(camel_cased_word)
-    return camel_cased_word unless /[A-Z-]/.match?(camel_cased_word)
-
-    word = camel_cased_word.gsub(/([A-Z\d]+)([A-Z][a-z])/, '\1_\2')
-    word.gsub!(/([a-z\d])([A-Z])/, '\1_\2')
-    word.tr!('-', '_')
-    word.downcase!
-    word
-  end
-end
-
-# Typescript < Codegen
-# Automatic Code generation overrides for TypeScript programming language
-class TypeScript < Codegen
-  def initialize(paths)
-    language_type_by_schema_type = {
-      'integer' => 'number',
-      'string' => 'string',
-      'boolean' => 'boolean'
-    }
-
-    super(paths, language_type_by_schema_type)
-  end
-
-  def array_type_for(type_name)
-    "readonly #{type_name}[]"
-  end
-end
-
-# Cpp < Codegen
-# Automatic Code generation overrides for C++ programming language
-class Cpp < Codegen
-  def initialize(paths)
-    language_type_by_schema_type = {
-      'integer' => 'std::size_t',
-      'string' => 'std::string',
-      'boolean' => 'bool'
-    }
-
-    super(paths, language_type_by_schema_type)
-  end
-
-  def array_type_for(type_name)
-    "std::vector<#{type_name}>"
-  end
-
-  def format_description(raw_description, indent_string: '')
-    return '' if raw_description.nil?
-
-    raw_description
-      .split("\n")
-      .map(&:strip)
-      .filter { |line| line != '*' }
-      .map { |line| "// #{line}".rstrip }
-      .join("\n#{indent_string}")
-  end
-end
-
-# Java < Codegen
-# Automatic Code generation overrides for Java programming language
-class Java < Codegen
-  def initialize(paths)
-    language_type_by_schema_type = {
-      'integer' => 'Long',
-      'string' => 'String',
-      'boolean' => 'Boolean'
-    }
-
-    super(paths, language_type_by_schema_type)
-  end
-
-  def array_type_for(type_name)
-    "java.util.List<#{type_name}>"
-  end
-
-  def format_description(raw_description, indent_string: '')
-    return '' if raw_description.nil?
-
-    raw_description
-      .split("\n")
-      .map(&:strip)
-      .filter { |line| line != '*' }
-      .map { |line| " * #{line}".rstrip }
-      .join("\n#{indent_string}")
-  end
-end
-
-# Perl < Codegen
-# Automatic Code generation overrides for Perl programming language
-class Perl < Codegen
-  def initialize(paths)
-    language_type_by_schema_type = {
-      'integer' => 'number',
-      'string' => 'string',
-      'boolean' => 'boolean'
-    }
-
-    super(paths, language_type_by_schema_type)
-  end
-
-  def array_type_for(type_name)
-    "[]#{type_name}"
-  end
-
-  def property_type_from_ref(ref)
-    "Cucumber::Messages::#{class_name(ref)}"
-  end
-
-  def property_type_from_enum(_enum)
-    ''
-  end
-
-  def format_description(raw_description)
-    return '' if raw_description.nil?
-
-    raw_description
-  end
-
-  private
-
-  def default_value(parent_type_name, property_name, property)
-    if property['type'] == 'string'
-      if property['enum']
-        "#{property_name.upcase}_#{enum_constant(property['enum'][0])}"
-      else
-        "''"
-      end
-    elsif property['type'] == 'boolean'
-      "''" # an empty string renders evaluates to false
-    elsif property['$ref']
-      type = type_for(parent_type_name, nil, property)
-      "#{type}->new()"
-    else
-      super(parent_type_name, property_name, property)
-    end
-  end
-end
-
-# Ruby < Codegen
-# Automatic Code generation overrides for Ruby programming language
-class Ruby < Codegen
-  def initialize(paths)
-    language_type_by_schema_type = {
-      'integer' => 'number',
-      'string' => 'string',
-      'boolean' => 'boolean'
-    }
-
-    super(paths, language_type_by_schema_type)
-  end
-
-  def array_type_for(_type_name)
-    '[]'
-  end
-
-  def format_description(raw_description, indent_string: '    ')
-    return '##' if raw_description.nil?
-
-    raw_description
-      .split("\n")
-      .map { |description_line| line_as_comment(description_line) }
-      .push('##')
-      .join("\n#{indent_string}")
-  end
-
-  private
-
-  def default_value(parent_type_name, property_name, property)
-    if property['type'] == 'string'
-      if property['enum']
-        enum_type_name = type_for(parent_type_name, property_name, property)
-        "#{enum_type_name}::#{enum_constant(property['enum'][0])}"
-      else
-        "''"
-      end
-    elsif property['$ref']
-      type = type_for(parent_type_name, nil, property)
-      "#{type}.new"
-    else
-      super(parent_type_name, property_name, property)
-    end
-  end
-
-  def line_as_comment(line)
-    if line.empty?
-      '#'
-    else
-      "# #{line}"
-    end
-  end
-end
-
-# Go < Codegen
-# Automatic Code generation overrides for Go programming language
-class Go < Codegen
-  def initialize(paths)
-    language_type_by_schema_type = {
-      'integer' => 'int64',
-      'string' => 'string',
-      'boolean' => 'bool'
-    }
-    super(paths, language_type_by_schema_type)
-  end
-
-  def property_type_from_ref(ref)
-    "*#{class_name(ref)}"
-  end
-
-  def array_type_for(type_name)
-    "[]#{type_name}"
-  end
-end
-
-# Markdown < Codegen
-# Automatic Code generation overrides for 'Cucumber with Markdown' programming language
-# NB: This is based on the cucumber-js implementation
-class Markdown < Codegen
-  def initialize(paths)
-    language_type_by_schema_type = {
-      'integer' => 'integer',
-      'string' => 'string',
-      'boolean' => 'boolean'
-    }
-    super(paths, language_type_by_schema_type)
-  end
-
-  def property_type_from_ref(ref)
-    "[#{class_name(ref)}](##{class_name(ref).downcase})"
-  end
-
-  def property_type_from_enum(enum)
-    "[#{enum}](##{enum.downcase})"
-  end
-
-  def array_type_for(type_name)
-    "#{type_name}[]"
-  end
-end
-
-# Php < Codegen
-# Automatic Code generation overrides for Php programming language
-class Php < Codegen
-  def initialize(paths)
-    language_type_by_schema_type = {
-      'string' => 'string',
-      'integer' => 'int',
-      'boolean' => 'bool'
-    }
-    super(paths, language_type_by_schema_type)
-  end
-
-  def format_description(raw_description, indent_string: '        ')
-    return '' if raw_description.nil?
-
-    raw_description
-      .split("\n")
-      .map(&:strip)
-      .filter { |line| line != '*' }
-      .map { |line| " * #{line}".rstrip }
-      .join("\n#{indent_string}")
-  end
-
-  def array_type_for(_type_name)
-    'array'
-  end
-
-  def enum_name(parent_type_name, property_name, enum)
-    enum_type_name = "#{class_name(parent_type_name)}\\#{capitalize(property_name)}"
-    @enum_set.add({ name: enum_type_name, values: enum })
-    enum_type_name
-  end
-
-  def array_contents_type(parent_type_name, _property_name, property)
-    type_for(parent_type_name, nil, property['items'])
-  end
-
-  def is_nullable(property_name, schema)
-    !(schema['required'] || []).index(property_name)
-  end
-
-  def is_scalar(property)
-    property.has_key?('type') && @language_type_by_schema_type.has_key?(property['type'])
-  end
-
-  def scalar_type_for(property)
-    unless @language_type_by_schema_type[property['type']]
-      raise "No type mapping for JSONSchema type #{type}. Schema:\n#{JSON.pretty_generate(property)}"
-    end
-
-    @language_type_by_schema_type[property['type']]
-  end
-
-  def constructor_for(parent_type, property, property_name, schema, arr_name)
-    constr = non_nullable_constructor_for(parent_type, property, property_name, schema, arr_name)
-
-    is_nullable(property_name, schema) ? "isset($#{arr_name}['#{property_name}']) ? #{constr} : null" : constr
-  end
-
-  def non_nullable_constructor_for(parent_type, property, property_name, schema, arr_name)
-    source = property_name.nil? ? arr_name : "#{arr_name}['#{property_name}']"
-    if is_scalar(property)
-      if property['enum']
-        "#{enum_name(parent_type, property_name, property['enum'])}::from((#{scalar_type_for(property)}) $#{source})"
-      else
-        "(#{scalar_type_for(property)}) $#{source}"
-      end
-    else
-      type = type_for(parent_type, property_name, property)
-      if type == 'array'
-        constructor = non_nullable_constructor_for(parent_type, property['items'], nil, schema, 'member')
-        member_type = (property['items']['type'] ? 'mixed' : 'array')
-        "array_values(array_map(fn (#{member_type} $member) => #{constructor}, $#{source}))"
-      else
-        "#{type_for(parent_type, property_name, property)}::fromArray($#{source})"
-      end
-    end
-  end
-
-  def default_enum(enum_type_name, property)
-    "#{enum_type_name}::#{enum_constant(property['enum'][0])}"
-  end
-
-  private
-
-  def default_value(class_name, property_name, property, schema)
-    return 'null' if is_nullable(property_name, schema)
-
-    super(class_name, property_name, property)
-  end
-end
-
 clazz = Object.const_get(ARGV[0])
 path = ARGV[1]
 paths = File.file?(path) ? [path] : Dir["#{path}/*.json"]
