+++ codegen/codegen/typescript.rb
@@ -1,0 +1,21 @@
+# frozen_string_literal: true
+
+module Codegen
+  # CodeGen::Typescript
+  # Automatic Code generation overrides for TypeScript programming language
+  class TypeScript < Base
+    def initialize(paths)
+      language_type_by_schema_type = {
+        'integer' => 'number',
+        'string' => 'string',
+        'boolean' => 'boolean'
+      }
+
+      super(paths, language_type_by_schema_type)
+    end
+
+    def array_type_for(type_name)
+      "readonly #{type_name}[]"
+    end
+  end
+end
